:recycle: 編集ページの構造をロジックと表示部分に分割
