+++ src/client/containers/contributes/Edit/index.tsx
@@ -1,28 +1,7 @@
-import { useEffect, useState } from "react";
-import PAGES from "@constants/pages";
-import SingleLineWideTemplate from "@components/templates/SingleLineWideTemplate";
-import MarkdownRenderer from "@components/atoms/MarkdownRenderer";
-import { ContributeInterface } from "../../../..//client/interface/contributes";
-import useExclusiveControl from "src/client/hooks/useExclusiveControl";
-import { useRouter } from "next/navigation";
-import styles from "./styles.module.scss";
-import { useUpdateEffect } from "src/client/hooks/useUpdateEffect";
 import { NextPage } from "next";
-import Textarea from "@components/atoms/Textarea";
-import TextForm from "@components/atoms/TextForm";
-import FloatButton from "@components/atoms/Buttons/FloatButton";
-import useContributeState from "src/client/state/contributes/contribute";
-import Button from "@components/atoms/Buttons";
-import Modal from "@components/molecules/Modal";
-import { CONTRIBUTE_STATUS } from "@server/domain/entity/contribute";
-import { FLASH_TYPE, useFlashMessage } from "@components/atoms/Flash";
-import Image from "next/image";
 import EditContributeSkeleton from "./Skeleton";
-
-// ここに置くべきではなさそう
-const AUTO_SAVE_INTERVAL = 10000; // 自動保存の間隔 (単位:ms)
-
-const contributeInterface = new ContributeInterface();
+import { useContributeEditor, usePublishModal } from "./hooks";
+import ContributeEditPresenter from "./presenter";
 
 type IContributeEditProps = {
   params: {
@@ -33,163 +12,29 @@ type IContributeEditProps = {
 const EditContribute: NextPage<IContributeEditProps> = ({
   params,
 }: IContributeEditProps) => {
-  const router = useRouter();
-  const exclude = useExclusiveControl();
-  const { contribute, setContribute } = useContributeState();
-  const [saveTimer, setSaveTimer] = useState<NodeJS.Timeout | undefined>(
-    undefined,
-  );
-  const [isOpenStatusModal, setIsOpenStatusModal] = useState<boolean>(false);
-  const { showFlashMessage } = useFlashMessage();
-  const { identityCode } = params;
-  const canPublish = contribute?.status === CONTRIBUTE_STATUS.DRAFT;
+  const {
+    contribute,
+    setTitle,
+    setContent,
+    updateContributeStatus,
+    canPublish,
+  } = useContributeEditor(params.identityCode);
+  const { isOpenStatusModal, openStatusModal, closeStatusModal } =
+    usePublishModal();
 
-  useEffect(() => {
-    if (identityCode) {
-      fetchContribute(identityCode as string);
-    }
-  }, [identityCode]);
-
-  useUpdateEffect(() => {
-    clearTimeout(saveTimer);
-    setSaveTimer(setTimeout(updateContribute, AUTO_SAVE_INTERVAL));
-  }, [contribute]);
-
-  const fetchContribute = async (identityCode: string) => {
-    try {
-      const contribute = await contributeInterface.getContribute(identityCode);
-      setContribute(contribute);
-    } catch (e) {
-      // TODO: エラー処理
-    }
-  };
-
-  const setTitle = (title: string) => {
-    if (!contribute) {
-      return;
-    }
-    setContribute({ ...contribute, title });
-  };
-
-  const setContent = (content: string) => {
-    if (!contribute) {
-      return;
-    }
-    setContribute({ ...contribute, content });
-  };
-
-  const updateContribute = async () => {
-    if (!contribute?.identityCode) {
-      return;
-    }
-
-    exclude(async () => {
-      await contributeInterface.updateContribute(contribute);
-    }, 500);
-  };
-
-  const updateContributeStatus = async () => {
-    if (!contribute?.identityCode) {
-      return;
-    }
-
-    exclude(async () => {
-      setContribute({
-        ...contribute,
-        status: CONTRIBUTE_STATUS.PUBLISHED,
-      });
-      const updatedContribute =
-        await contributeInterface.updateContributeStatus({
-          ...contribute,
-          status: CONTRIBUTE_STATUS.PUBLISHED,
-        });
-      if (!updatedContribute) {
-        showFlashMessage(
-          "記事の公開に失敗しました。もう1度お試しいただくか、カスタマーサポートまでご連絡ください。",
-          FLASH_TYPE.ERROR,
-        );
-        closeStatusModal();
-      } else {
-        showFlashMessage(`記事を公開しました。`, FLASH_TYPE.SUCCESS);
-        router.push(PAGES.HOME.PATH);
-      }
-    }, 500);
-  };
-
-  const openStatusModal = () => {
-    setIsOpenStatusModal(true);
-  };
-
-  const closeStatusModal = () => {
-    setIsOpenStatusModal(false);
-  };
-
-  return (
-    <SingleLineWideTemplate
-      pageTitle={PAGES.CONTRIBUTES_EDIT.TITLE}
-      pageDescription={PAGES.CONTRIBUTES_EDIT.DESCRIPTION}
-    >
-      {!contribute ? (
-        <EditContributeSkeleton />
-      ) : (
-        <section>
-          <TextForm
-            content={contribute?.title || ""}
-            placeholder="まずはタイトルを入力しましょう"
-            onChange={setTitle}
-          />
-          <div className={styles.contentContainer}>
-            <div className={styles.writeContainer}>
-              <Textarea
-                content={contribute?.content || ""}
-                placeholder="本文を入力しましょう。マークダウン記法に対応しています。"
-                onChange={setContent}
-              />
-            </div>
-            <div className={styles.previewContainer}>
-              <div className={styles.preview}>
-                <MarkdownRenderer
-                  content={
-                    contribute?.content || "ここにプレビューが表示されます。"
-                  }
-                />
-              </div>
-            </div>
-          </div>
-          {canPublish && (
-            <div className={styles.floatButtonContainer}>
-              <FloatButton
-                text={
-                  <Image
-                    src="/icons/earth.svg"
-                    alt="公開"
-                    width={25}
-                    height={25}
-                    color="white"
-                  />
-                }
-                onClick={openStatusModal}
-              />
-            </div>
-          )}
-          {isOpenStatusModal && (
-            <Modal onCancel={closeStatusModal}>
-              <h2 className={styles.modalTitle}>この記事を公開しますか？</h2>
-              <p className={styles.modalContent}>
-                この記事を公開すると、全てのユーザーが閲覧できるようになります。
-              </p>
-              <div className={styles.modalButtonContainer}>
-                <Button
-                  text="公開する"
-                  onClick={updateContributeStatus}
-                  type="main"
-                />
-              </div>
-            </Modal>
-          )}
-        </section>
-      )}
-    </SingleLineWideTemplate>
+  return contribute ? (
+    <ContributeEditPresenter
+      contribute={contribute}
+      setTitle={setTitle}
+      setContent={setContent}
+      updateContributeStatus={updateContributeStatus}
+      canPublish={canPublish}
+      isOpenStatusModal={isOpenStatusModal}
+      openStatusModal={openStatusModal}
+      closeStatusModal={closeStatusModal}
+    />
+  ) : (
+    <EditContributeSkeleton />
   );
 };
